db update

+++ posts/views.py
@@ -130,18 +130,19 @@ def edit_post(request, pk):
         return redirect('homepage:home')
     if request.method == 'POST' and form.is_valid():
         update_attrs(curr_post, **form.cleaned_data)
-
-        curr_post_gallery = PostsGallery.objects.post_gallery(item_id=pk)
-        for img in curr_post_gallery:
-            img.delete()
-
         files = request.FILES.getlist('gallery')
-        for img in files:
-            new_gallery = PostsGallery.objects.create(
-                item=curr_post,
-                upload=img,
-            )
-            new_gallery.save()
+
+        if files:
+            curr_post_gallery = PostsGallery.objects.post_gallery(item_id=pk)
+            for img in curr_post_gallery:
+                img.delete()
+
+            for img in files:
+                new_gallery = PostsGallery.objects.create(
+                    item=curr_post,
+                    upload=img,
+                )
+                new_gallery.save()
 
         messages.success(request, 'Пост был успешно обновлен')
         return redirect('users:profile')
